Add workflow to check for VERSION edits (#8)
* Add workflow to check for VERSION edits

Developers should not edit the VERSION file.

* Add checkout to workflow

Can't check the version if there's no code.

* Adjust fetch depth

* Add origin to branch name

* Specify VERSION file in diff

It really helps to specify which file you want to diff.

* Remove pull_request trigger

It seems that the pull_request trigger is unnecessary.

* Edit the VERSION file

Let's see if the workflow catches it.

* Fix version check error message

I needed single-quotes to prevent variable expansion.

* Fix quotation marks

* Restore VERSION

* Edit VERSION file

Test what happens when I edit the file with the pull request closed.

* Restore VERSION

* Test fetch depth 2

* Fetch depth needs to be 0

* Update VERSION again

* Update test.txt

* Restore text.txt

* Restore VERSION

* Add debugging info

Let's try to checkout only the needed branch, rather than the entire
commit history.

* Switch to pull request trigger

* Update version check

* Update version check

* Update version check

* Restore VERSION

* Update version check

* Edit VERSION

* Restrict comment length

* Try no git switch

* Restore VERSION

* Remove git switch

* See what happens when I remove main branch

* Restore main branch checkout

Yup, it didn't work.

* Edit failure message

* Fix typo

* Update VERSION

* Restore VERSION